- Clean up logic around PlayState refs, removing Game reference and moving to a simpler static singleton - Basic support quests to be published to server - Support playing a server-published quest - Comms form to always be top/right over game window

+++ SrvSurvey/quests/DefQuest.cs
@@ -4,7 +4,7 @@
 public class DefQuest
 {
     public required string id;
-    public required float ver;
+    public required double ver;
     public required string publisher;
     public required string title;
     public string? desc;
